refactor: consolidate isConnected/isLedgerReadOnly check into useIsWalletConnected hook
Creates a reusable useIsWalletConnected hook that encapsulates the repetitive
pattern of checking both wallet connection state and Ledger read-only mode.

Refactors 18 files to use the new hook, eliminating the boilerplate 5-line
pattern repeated across:
- Fox components: FoxTokenBalances, FoxGovernance, RFOXSection, FoxFarming, FoxHeader
- Dashboard: DashboardHeaderTop, AccountTable, Dashboard
- Trade/TX: TradeAssetInput, TransactionHistoryContent, WalletBalanceChange
- Common: AccountDropdown, AssetRow, GroupedAssetRow, AssetActions, ActionCenter
- Other: TCY page, YieldAvailableToDeposit

Closes SS-5359
