+++ src/pages/Fox/components/FoxGovernance.tsx
@@ -24,15 +24,13 @@ import { FoxGovernanceProposal } from './FoxGovernanceProposal'
 import { Amount } from '@/components/Amount/Amount'
 import { Text } from '@/components/Text'
 import type { TextPropTypes } from '@/components/Text/Text'
-import { KeyManager } from '@/context/WalletProvider/KeyManager'
 import { useFeatureFlag } from '@/hooks/useFeatureFlag/useFeatureFlag'
-import { useWallet } from '@/hooks/useWallet/useWallet'
+import { useIsWalletConnected } from '@/hooks/useIsWalletConnected/useIsWalletConnected'
 import {
   selectIsSnapshotApiQueriesPending,
   selectVotingPowerOrZero,
 } from '@/state/apis/snapshot/selectors'
 import { snapshotApi, useGetProposalsQuery } from '@/state/apis/snapshot/snapshot'
-import { selectWalletType } from '@/state/slices/localWalletSlice/selectors'
 import { selectAssetById, selectWalletAccountIds } from '@/state/slices/selectors'
 import { useAppDispatch, useAppSelector } from '@/state/store'
 
@@ -56,13 +54,7 @@ const widthMdAuto = { base: 'full', md: 'auto' }
 const tabListPaddingLeft = { base: 6, md: 0 }
 
 export const FoxGovernance = () => {
-  const {
-    state: { isConnected: isWalletConnected },
-  } = useWallet()
-  const walletType = useAppSelector(selectWalletType)
-  const isLedgerReadOnlyEnabled = useFeatureFlag('LedgerReadOnly')
-  const isLedgerReadOnly = isLedgerReadOnlyEnabled && walletType === KeyManager.Ledger
-  const isConnected = isWalletConnected || isLedgerReadOnly
+  const isConnected = useIsWalletConnected()
 
   const translate = useTranslate()
   const isFoxGovernanceEnabled = useFeatureFlag('FoxPageGovernance')
