+++ src/pages/Yields/components/YieldAvailableToDeposit.tsx
@@ -16,12 +16,9 @@ import { useTranslate } from 'react-polyglot'
 
 import { Amount } from '@/components/Amount/Amount'
 import { SwapperModal } from '@/components/SwapperModal/SwapperModal'
-import { KeyManager } from '@/context/WalletProvider/KeyManager'
-import { useFeatureFlag } from '@/hooks/useFeatureFlag/useFeatureFlag'
-import { useWallet } from '@/hooks/useWallet/useWallet'
+import { useIsWalletConnected } from '@/hooks/useIsWalletConnected/useIsWalletConnected'
 import { bnOrZero } from '@/lib/bignumber/bignumber'
 import type { AugmentedYieldDto } from '@/lib/yieldxyz/types'
-import { selectWalletType } from '@/state/slices/localWalletSlice/selectors'
 import { selectPortfolioCryptoBalanceByFilter } from '@/state/slices/selectors'
 import { useAppSelector } from '@/state/store'
 
@@ -34,18 +31,7 @@ export const YieldAvailableToDeposit = memo(
   ({ yieldItem, inputTokenMarketData }: YieldAvailableToDepositProps) => {
     const translate = useTranslate()
     const [isSwapperModalOpen, setIsSwapperModalOpen] = useState(false)
-    const {
-      state: { isConnected },
-    } = useWallet()
-    const isLedgerReadOnlyEnabled = useFeatureFlag('LedgerReadOnly')
-    const walletType = useAppSelector(selectWalletType)
-    const isLedgerReadOnly = isLedgerReadOnlyEnabled && walletType === KeyManager.Ledger
-
-    // Either wallet is physically connected, or it's a Ledger in read-only mode
-    const hasWallet = useMemo(
-      () => isConnected || isLedgerReadOnly,
-      [isConnected, isLedgerReadOnly],
-    )
+    const hasWallet = useIsWalletConnected()
 
     const inputToken = yieldItem.inputTokens[0]
     const inputTokenSymbol = inputToken?.symbol ?? yieldItem.token.symbol
